+++ code/packages/rust/vault-pm-application/src/lib.rs
@@ -9,9 +9,15 @@
 
 mod codec;
 mod crypto;
+mod verifier;
 
-pub use codec::{decode_item_revision, encode_item_revision, CatalogV1, LocalSecretV1};
+pub use codec::{
+    decode_device_certificate, decode_item_revision, decode_signed_commit,
+    encode_device_certificate, encode_item_revision, encode_signed_commit, CatalogV1,
+    LocalSecretV1,
+};
 pub use crypto::{open_object, seal_object, ObjectKind, ObjectRandomness, V1Keys};
+pub use verifier::V1SingleDeviceVerifier;
 
 use core::fmt::{self, Debug, Display, Formatter};